requirements.txt
opencv-python==4.12.0.88
mediapipe==0.10.21
numpy==1.26.4
pyautogui==0.9.54
Pillow==11.3.0